add auth & social layer: request friends, lists of incoming & outcoming requests

+++ src/main/java/ru/technocracy/echoserver/models/notedevent/NotedEventPK.java
@@ -15,5 +15,5 @@ public class NotedEventPK implements Serializable {
 
     private Long userId;
 
-    private Long eventId;
+    private int eventId;
 }
